Use pip install routines in pip sync
