+++ crates/uv/src/commands/pip/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod compile;
 pub(crate) mod freeze;
 pub(crate) mod install;
 pub(crate) mod list;
+mod operations;
 pub(crate) mod show;
 pub(crate) mod sync;
 pub(crate) mod uninstall;
